Add in testing

+++ bin/prebuild.sh
@@ -3,7 +3,11 @@
 taxids=$1
 
 rm -rf /tmp/interpro
-rm have_latest_interpro
+
+if [ -e have_latest_interpro ]; then
+	rm have_latest_interpro
+fi
+
 rm *version.txt
 
 # Check for the latest version of InterPro from the source servers
@@ -19,9 +23,15 @@ echo "Checking for presence of ${taxids%,*} as extracted InterPro"
 echo "Checking S3 path s3:::${BUILD_OUTPUT_BUCKET}/${BUILD_OUTPUT_PREFIX}/interpro/InterPro-${interpro_version}-${taxids%,*}.tsv"
 
 checkversion --fail-on-match \
-			 --print \
 			 --s3path "s3:::${BUILD_OUTPUT_BUCKET}/${BUILD_OUTPUT_PREFIX}/interpro/InterPro-${interpro_version}-${taxids%,*}.tsv" \
-			 --static "$interpro_version" && echo "No existing file" || touch 'have_latest_interpro' && echo "We have an existing InterPro build for this release"
+			 --static "$interpro_version"
+
+have_existing_interpro=$?
+if [ $have_existing_interpro -gt 0 ]; then
+	touch 'have_latest_interpro' && echo "We have an existing InterPro build for this release"
+else
+	echo "No existing file"
+fi
 
 # Grab the files that have already been parsed for InterPro
 
@@ -52,6 +62,23 @@ glycodomain_version=$(<"glycodomain_version.txt")
 
 echo "Checking for domains with version InterPro-${interpro_version}-Glycodomain-${glycodomain_version}"
 
+exit_code=1
+
 for taxid in ${taxids//,/ }; do
-	testversion "glycodomain_${taxid}.json" --static "Interpro-${interpro_version}-Glycodomain-${glycodomain_version}"
+	echo "Checking existence of Glycodomain json for $taxid"
+	testversion_skip_exit "glycodomain_${taxid}.json" --static "Interpro-${interpro_version}-Glycodomain-${glycodomain_version}"
+	retcode=$?
+	if [ $retcode -ne 0 ]; then
+		echo "Existing Glycodomain json for $taxid"
+	else
+		echo "No existing Glycodomain json for $taxid"
+		exit_code=0
+	fi
 done
+
+if [ $exit_code -eq 0 ]; then
+	true
+else
+	echo "Glycodomain files are up to date"
+	false
+fi
